feat: use already calculated exc_text
The logging module sets the traceback string to `record.exc_text` and
doesn't caluclate it again.
This allows also to log exceptions, when we are not in the stack-context
anymore and have no `exc_info` tuple.
For example, in some cases, only the textual representation of a
traceback is stored because the traceback object would be a memory leak
and prevent garbage collection.

Use case is therefor:
`formatter.format(record)` where `exc_text` was already created by
another earlier formatter or filter.