fix: always quote control chars
None of the invisible ASCII Control chars 0 - 32 were replaced.
This PR replaces them with a unicode representation, the same as go-logfmt does it.
e.g. `\x00` → `"\u0000"`

This is security critical, as currently `\r` and other things would be part of the raw log.
It would e.g. allow via ANSI escape sequences to set information in a linux terminal, like the window name/title.